requirements.txt
python-telegram-bot==13.10
requests==2.27.1
coverage==6.2
